This logic was never exposed in the /export route
