+++ creator-node/docker-compose/development.env
@@ -26,8 +26,8 @@ serviceCountry=
 
 # required when running with kubernetes, need to pass in the ip address of the instance that will run 
 # the ipfs node. should be done by default in audius manifests
-# ipfsClusterIP=
-# ipfsClusterPort=
+# ipfsClusterIP=127.0.0.1
+# ipfsClusterPort=4001
 
 # wallet information required on all environments
 delegateOwnerWallet=
